Add files via upload
Generating the Training and Testing CSV Files
To make the dataset easier to work with, we will create two CSV files. A trainval.csv and a test.csv file. They will contain two columns. One for the path to each image in the dataset and one for the corresponding label. For the labels, we consider 0 as ‘Crack Detected’ and 1 as ‘Crack Undetected’. 0 means image folder starting with ‘C’, as in ‘CW’, and 1 means image folder starting with ‘U’ as in ‘UW’.

The code to create this CSV file is present in the prepare_data.py file.

Although we will not go into the details of this code, the following block contains the entire code. Please go through it if you need to.