Add files via upload
The Utility and Helper Scripts
We need a few helper functions and a utility class to make our work easier throughout the project.

All of the code for this will be in the utils.py file.

Let’s start with the import statements and a function to save the trained PyTorch model.
The save_model function saves the trained model weights along with the number of epochs trained for, the optimizer state dictionary, and the loss function. This will help us resume training in the future if we want to.

Next, we have a simple function to save the accuracy and loss graphs to disk.