Update utils.py
The LRScheduler class simply calls the ReduceLROnPlateau scheduler from torch.optim.lr_scheduler. The only reason we are again initializing this inside a custom class is to abstract it away from the training script. Basically, this will reduce the learning rate by a certain factor, when the validation loss does not improve for a certain number of epochs controlled by the patience parameter. This will help us prevent overfitting when the model starts to learn as the training progresses.

The above are all the helper functions and utility classes that we need.